pyproject.toml
[project]
name = "piaic-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
